+++ nextstep-di/src/test/java/nextstep/di/factory/support/ReflectionUtilsTest.java
@@ -3,6 +3,7 @@ package nextstep.di.factory.support;
 import nextstep.di.factory.example.JdbcUserRepository;
 import nextstep.di.factory.example.MyQnaService;
 import nextstep.di.factory.exception.ReflectionUtilException;
+import nextstep.di.factory.util.ReflectionUtils;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
